// src/interfaces/text.ts
export interface INgxText {
  default?: string;
  _default?: string;
  button?: string;
  try_again?: string;
  replace?: string;
  reset?: string;
  error?: string;
}

export interface INgxErrorText {
  fileSize?: string;
  minWidth?: string;
  maxWidth?: string;
  minHeight?: string;
  maxHeight?: string;
  imageFormat?: string;
  fileType?: string;
}